refactor: reorganize settings module and remove deprecated files
- Consolidated language, theme, and username functionalities into dedicated service files for better modularity.
- Updated language and theme settings to utilize new service methods for improved maintainability.
- Removed obsolete test files and components related to username management and theme handling.
- Streamlined translation keys to focus on essential settings, enhancing clarity and reducing redundancy.
