+++ src/renderer/features/settings/i18n/en.ts
@@ -1,26 +1,10 @@
+import { setLanguageEn } from '../set-language/i18n'
+import { setThemeEn } from '../set-theme/i18n'
+import { setUsernameEn } from '../set-username/i18n'
+
 export default {
   title: 'Settings',
-  language: {
-    title: 'Language',
-    description: 'Choose the language in which Dojosphere should be displayed.'
-  },
-  theme: {
-    title: 'Color Scheme',
-    description: 'Determine whether Dojosphere is displayed in light or dark mode.',
-    tooltip: {
-      light: 'Light Mode',
-      dark: 'Dark Mode',
-      system: 'System Setting'
-    }
-  },
-  username: {
-    title: 'Username',
-    description: 'Your display name on this device.',
-    save: 'Save',
-    success: 'Username saved.',
-    error: {
-      empty: 'Username must not be empty.',
-      save: 'Username could not be saved.'
-    }
-  }
+  language: setLanguageEn,
+  theme: setThemeEn,
+  username: setUsernameEn
 }
