+++ src/renderer/features/settings/i18n/keys.ts
@@ -1,28 +1,5 @@
 const translationKeys = {
-  title: 'settings.title',
-  language: {
-    title: 'settings.language.title',
-    description: 'settings.language.description'
-  },
-  theme: {
-    title: 'settings.theme.title',
-    description: 'settings.theme.description',
-    tooltip: {
-      light: 'settings.theme.tooltip.light',
-      dark: 'settings.theme.tooltip.dark',
-      system: 'settings.theme.tooltip.system'
-    }
-  },
-  username: {
-    title: 'settings.username.title',
-    description: 'settings.username.description',
-    save: 'settings.username.save',
-    success: 'settings.username.success',
-    error: {
-      empty: 'settings.username.error.empty',
-      save: 'settings.username.error.save'
-    }
-  }
+  title: 'settings.title'
 }
 
 export default translationKeys
