+++ src/renderer/features/settings/set-language/i18n/de.ts
@@ -1,0 +1,4 @@
+export default {
+  title: 'Sprache',
+  description: 'Wähle die Sprache aus, in der Dojosphere angezeigt werden soll.'
+}
