+++ src/renderer/features/settings/set-language/i18n/en.ts
@@ -1,0 +1,4 @@
+export default {
+  title: 'Language',
+  description: 'Choose the language in which Dojosphere should be displayed.'
+}
