+++ src/renderer/features/settings/set-language/i18n/index.ts
@@ -1,0 +1,2 @@
+export { default as setLanguageDe } from './de'
+export { default as setLanguageEn } from './en'
