+++ src/renderer/features/settings/set-language/i18n/keys.ts
@@ -1,0 +1,6 @@
+const translationKeys = {
+  title: 'settings.language.title',
+  description: 'settings.language.description'
+}
+
+export default translationKeys
