+++ src/renderer/features/settings/set-language/index.ts
@@ -1,0 +1,3 @@
+export { getInitialLanguage } from './service/language-service'
+export { getLanguageFromStorage, setLanguageToStorage } from './service/language-storage'
+export { default as LanguageSwitch } from './ui/LanguageSwitch.vue'
