+++ src/renderer/features/settings/set-language/ui/LanguageSwitch.e2e.spec.ts
