+++ src/renderer/features/settings/set-language/ui/LanguageSwitch.stories.ts
@@ -3,7 +3,7 @@ import type { Meta, StoryObj } from '@storybook/vue3-vite'
 import LanguageSwitch from './LanguageSwitch.vue'
 
 const meta = {
-  title: 'Features/Settings/LanguageSwitch',
+  title: 'Features/Settings/SetLanguage/LanguageSwitch',
   component: LanguageSwitch,
   parameters: {
     layout: 'fullscreen'
