+++ src/renderer/features/settings/set-theme/i18n/de.ts
@@ -1,0 +1,9 @@
+export default {
+  title: 'Farbschema',
+  description: 'Bestimme, ob Dojosphere im hellen oder dunklen Modus angezeigt werden soll.',
+  tooltip: {
+    light: 'Heller Modus',
+    dark: 'Dunkler Modus',
+    system: 'Systemeinstellung'
+  }
+}
