+++ src/renderer/features/settings/set-theme/i18n/en.ts
@@ -1,0 +1,9 @@
+export default {
+  title: 'Color Scheme',
+  description: 'Determine whether Dojosphere is displayed in light or dark mode.',
+  tooltip: {
+    light: 'Light Mode',
+    dark: 'Dark Mode',
+    system: 'System Setting'
+  }
+}
