+++ src/renderer/features/settings/set-theme/i18n/index.ts
@@ -1,0 +1,2 @@
+export { default as setThemeDe } from './de'
+export { default as setThemeEn } from './en'
