+++ src/renderer/features/settings/set-theme/i18n/keys.ts
@@ -1,0 +1,11 @@
+const translationKeys = {
+  title: 'settings.theme.title',
+  description: 'settings.theme.description',
+  tooltip: {
+    light: 'settings.theme.tooltip.light',
+    dark: 'settings.theme.tooltip.dark',
+    system: 'settings.theme.tooltip.system'
+  }
+}
+
+export default translationKeys
