+++ src/renderer/features/settings/set-theme/index.ts
@@ -1,0 +1,3 @@
+export { getInitialTheme } from './service/theme-service'
+export { getThemeFromStorage, setThemeToStorage } from './service/theme-storage'
+export { default as ThemeToggle } from './ui/ThemeToggle.vue'
