+++ src/renderer/features/settings/set-theme/service/theme-service.test.ts
