+++ src/renderer/features/settings/set-theme/service/theme-service.ts
