+++ src/renderer/features/settings/set-theme/service/theme-storage.ts
