+++ src/renderer/features/settings/set-theme/ui/ThemeToggle.e2e.spec.ts
