+++ src/renderer/features/settings/set-username/i18n/de.ts
@@ -1,0 +1,10 @@
+export default {
+  title: 'Benutzername',
+  description: 'Dein Anzeigename auf diesem Gerät.',
+  save: 'Speichern',
+  success: 'Benutzername gespeichert.',
+  error: {
+    empty: 'Der Benutzername darf nicht leer sein.',
+    save: 'Der Benutzername konnte nicht gespeichert werden.'
+  }
+}
