+++ src/renderer/features/settings/set-username/i18n/en.ts
@@ -1,0 +1,10 @@
+export default {
+  title: 'Username',
+  description: 'Your display name on this device.',
+  save: 'Save',
+  success: 'Username saved.',
+  error: {
+    empty: 'Username must not be empty.',
+    save: 'Username could not be saved.'
+  }
+}
