+++ src/renderer/features/settings/set-username/i18n/index.ts
@@ -1,0 +1,2 @@
+export { default as setUsernameDe } from './de'
+export { default as setUsernameEn } from './en'
