+++ src/renderer/features/settings/set-username/i18n/keys.ts
@@ -1,0 +1,12 @@
+const translationKeys = {
+  title: 'settings.username.title',
+  description: 'settings.username.description',
+  save: 'settings.username.save',
+  success: 'settings.username.success',
+  error: {
+    empty: 'settings.username.error.empty',
+    save: 'settings.username.error.save'
+  }
+}
+
+export default translationKeys
