+++ src/renderer/features/settings/set-username/index.ts
@@ -1,0 +1,1 @@
+export { default as UsernameEditor } from './ui/UsernameEditor.vue'
