+++ src/renderer/features/settings/set-username/service/ensure-local-session-for-username.test.ts
