+++ src/renderer/features/settings/set-username/service/update-display-name.test.ts
