+++ src/renderer/features/settings/set-username/ui/UsernameEditor.e2e.spec.ts
